+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- On macOS 27 and later, the MAC address of the interfaces was reported as 02:00:00:00:00:00 (kernel APIs now redact the link-layer address for non-root processes). It is now retrieved from SystemConfiguration when redacted.
 
 ## [1.2.10] - 2026-05-27
 ### Fixed